// degree37-2.0-api/src/api/common/enums/polymorphic-type.enum.ts
export enum PolymorphicType {
  CRM_ACCOUNTS = 'accounts', // CRM -> Accounts
  CRM_LOCATIONS = 'crm_locations', // CRM -> Locations
  CRM_CONTACTS_STAFF = 'staff', //  CRM -> Contacts -> Staff
  CRM_CONTACTS_VOLUNTEERS = 'crm_volunteer', // CRM -> Contacts -> Volunteer
  CRM_CONTACTS_DONORS = 'donors', //  CRM -> Contacts -> Donors
  CRM_DONOR_CENTERS = 'donor_centers', // CRM -> Donor Centers
  CRM_NON_COLLECTION_PROFILES = 'crm_non_collection_profiles', // CRM -> Non-Collection Profile
  OC_OPERATIONS_DRIVES = 'drives', //  OC -> Operations -> Drives
  OC_OPERATIONS_SESSIONS = 'sessions', //  OC -> Operations -> Sessions
  OC_OPERATIONS_NON_COLLECTION_EVENTS = 'oc_non_collection_events', //  OC -> Operations -> Non Collection Events
}